pyproject.toml
[project]
name = "claw-agent"
version = "0.1.0"
description = "A simple AI agent"
requires-python = ">=3.12"
dependencies = [
    # Web
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.0",
    "python-multipart==0.0.9",
    # LLM
    "langchain==0.3.7",
    "langchain-core==0.3.21",
    "langchain-openai==0.2.10",
    # MCP
    "mcp==1.3.0",
    "sse-starlette==2.1.3",
    # Database
    "elasticsearch==7.16.2",
    # HTTP & Networking
    "httpx==0.27.2",
    # Utils
    "loguru==0.7.2",
    "python-dotenv==1.0.1",
    "python-frontmatter==1.1.0",
]